requirements.txt
# ==============================================================================
# PriceScout - Production Dependencies
# ==============================================================================
# Last Updated: November 24, 2025
# Python Version: 3.11+ (tested on 3.11, 3.13)
# Azure Ready: Yes
#
# Install: pip install -r requirements.txt
# Post-install: playwright install chromium
# ==============================================================================

# === CORE FRAMEWORK ===
# PINNED: Streamlit 1.38.0 - versions 1.39+ have broken download_button behavior
# behind nginx proxy (files download with UUID names instead of proper filenames).
# Issue discovered: 2025-11-26. Test newer versions by checking if downloads
# work correctly with proper filenames (e.g., "daily_lineup_Theater_2025-11-26.xlsx")
# GitHub issue to track: https://github.com/streamlit/streamlit/issues (search "download filename proxy")
streamlit==1.38.0              # Web application framework - PINNED (see note above)
pandas>=2.0.0,<3.0.0           # Data manipulation and analysis
altair>=5.0.0,<6.0.0           # Declarative visualization (used in analysis_mode)

# === WEB SCRAPING ===
playwright==1.55.0             # Browser automation for scraping
beautifulsoup4>=4.12.0,<5.0.0  # HTML parsing
httpx>=0.27.0,<1.0.0           # Async HTTP client
requests>=2.31.0,<3.0.0        # Sync HTTP client (OMDb API)

# === DATA PROCESSING ===
thefuzz[speedup]>=0.22.0       # Fuzzy string matching for theater names
python-Levenshtein>=0.25.0     # Fast string matching (thefuzz speedup)
RapidFuzz>=3.0.0               # Alternative fuzzy matching

# === DATABASE ===
SQLAlchemy>=2.0.0,<3.0.0       # ORM and database abstraction
psycopg2-binary>=2.9.0,<3.0.0  # PostgreSQL adapter (Azure PostgreSQL)
pyodbc>=5.0.0,<6.0.0           # MSSQL/Azure SQL adapter (requires ODBC Driver 18)
alembic>=1.12.0,<2.0.0         # Database migrations

# === SECURITY ===
bcrypt>=4.0.0,<6.0.0           # Password hashing (industry standard)
python-jose[cryptography]>=3.3.0 # JWT library for API authentication


# === AZURE CLOUD ===
azure-identity>=1.15.0,<2.0.0  # Azure authentication (Managed Identity)
azure-keyvault-secrets>=4.7.0,<5.0.0  # Azure Key Vault integration
opentelemetry-distro>=0.40b0    # OpenTelemetry distribution for Azure Monitor
azure-servicebus>=7.11.0       # Azure Service Bus for messaging
opentelemetry-instrumentation-fastapi>=0.40b0 # FastAPI instrumentation for OpenTelemetry
msal>=1.24.0                     # Microsoft Authentication Library for Entra ID

# === DATE/TIME ===
pytz>=2024.1                   # Timezone handling
python-dateutil>=2.8.0,<3.0.0  # Date parsing utilities

# === FILE HANDLING ===
openpyxl>=3.1.0,<4.0.0         # Excel read support (.xlsx)
xlsxwriter>=3.1.0,<4.0.0       # Excel write support (better formatting)

# === ENVIRONMENT ===
python-dotenv>=1.0.0,<2.0.0    # Environment variable management

# === API (Alpha) ===
fastapi>=0.111.0,<1.0.0        # REST API framework for programmatic access
uvicorn[standard]>=0.29.0,<1.0 # ASGI server for FastAPI
pydantic>=2.0.0,<3.0.0         # Data validation for API schemas
python-multipart>=0.0.9        # Form data parsing for FastAPI

# ==============================================================================
# TESTING (Dev/CI only - can be moved to requirements-dev.txt)
# ==============================================================================
pytest>=8.0.0,<9.0.0           # Test framework
pytest-asyncio>=0.23.0,<2.0.0  # Async test support
pytest-cov>=4.0.0,<8.0.0       # Coverage reporting
pytest-mock>=3.12.0,<4.0.0     # Mocking utilities

# ==============================================================================
# POST-INSTALLATION STEPS
# ==============================================================================
# After pip install, run:
#   playwright install chromium
#
# For Azure deployment, browsers are installed in Dockerfile.
# ==============================================================================
